requirements.txt
alabaster==0.7.10
appdirs==1.4.0
Babel==2.4.0
colorama==0.3.7
Django==1.10.5
django-cors-headers==2.0.2
django-pyodbc-azure==1.10.4.0
djangorestframework==3.5.3
docutils==0.13.1
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.12.0
packaging==16.8
pandas==0.19.2
Pygments==2.2.0
pyodbc==4.0.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.5
xlrd==1.0.0
